fix rating int value

+++ reviews/models.py
@@ -8,7 +8,7 @@ from profiles.models import UserProfile
 
 class Review(models.Model):
 
-    rating = models.DecimalField(max_digits=6, decimal_places=2)
+    rating = models.IntegerField()
     userid = models.ForeignKey(UserProfile, on_delete=models.CASCADE)
     product = models.ForeignKey(Product, on_delete=models.CASCADE)
     review_title = models.CharField(max_length=90)
